frontend/src/data/usecases/users/remote-user-reset-password/remote-user-reset-password.spec.ts
import faker from 'faker'
import {HttpClientSpy, mockErrorsDetails} from '../../../../data/test'
import {DetailError} from '../../../../domain/errors'
import {StatusCode} from '../../../../domain/protocols/http'
import {RemoteUserResetPassword} from './remote-user-reset-password'

type SutTypes = {
  sut: RemoteUserResetPassword
  httpClientSpy: HttpClientSpy
}

const makeSut = (url: string = faker.internet.url()): SutTypes => {
  const httpClientSpy = new HttpClientSpy()
  const sut = new RemoteUserResetPassword(url, httpClientSpy)

  return {
    sut,
    httpClientSpy,
  }
}

describe('RemoteUserResetPassword', () => {
  it('Should call HttpClient with correct values', async () => {
    const url = faker.internet.url()
    const {sut, httpClientSpy} = makeSut(url)
    httpClientSpy.response = {
      statusCode: StatusCode.ok,
    }

    await sut.resetPassword()

    expect(httpClientSpy.url).toEqual(url)
    expect(httpClientSpy.method).toBe('get')
  })

  it('Should throw UnprocessableEntityError if HttpClient returns 500', async () => {
    const {sut, httpClientSpy} = makeSut()
    const errorsDetails = mockErrorsDetails()
    httpClientSpy.response = {
      statusCode: StatusCode.serverError,
    }

    const promise = sut.resetPassword()

    await expect(promise).rejects.toThrow(new DetailError(errorsDetails))
  })
})
